+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/handlers/WebServiceSelectorRequestHandlerSpec.groovy
@@ -114,6 +114,7 @@ class WebServiceSelectorRequestHandlerSpec extends Specification {
                 Collections.emptyList(),
                 Collections.emptyList(),
                 Collections.emptyList(),
+                null,
                 null
         )
         Integer timeout = 20000